Feat: Criado DTO de todas as classes

+++ Back/src/ProEventos.API/Controllers/EventosController.cs
@@ -9,6 +9,7 @@ using ProEventos.Domain;
 using ProEventos.Persistence.Contextos;
 using ProEventos.Application.Contratos;
 using Microsoft.AspNetCore.Http;
+using ProEventos.Application.Dtos;
 
 namespace ProEventos.API.Controllers
 {
@@ -29,7 +30,7 @@ namespace ProEventos.API.Controllers
             try
             {
                 var eventos = await _eventoService.GetAllEventosAsync(true);
-                if (eventos == null) return NotFound("Nenhum evento encontrado.");
+                if (eventos == null) return NoContent();
 
                 return Ok(eventos);
             }
@@ -47,7 +48,7 @@ namespace ProEventos.API.Controllers
             try
             {
                 var evento = await _eventoService.GetEventoByIdAsync(id, true);
-                if (evento == null) return NotFound("Evento por Id não encontrado.");
+                if (evento == null) return NoContent();
 
                 return Ok(evento);
             }
@@ -65,7 +66,7 @@ namespace ProEventos.API.Controllers
             try
             {
                 var eventos = await _eventoService.GetEventoByTemaAsync(tema, true);
-                if (eventos == null) return NotFound("Eventos por tema não encontrados.");
+                if (eventos == null) return NoContent();
 
                 return Ok(eventos);
             }
@@ -78,7 +79,7 @@ namespace ProEventos.API.Controllers
         }
 
         [HttpPost]
-        public async Task<IActionResult> Post(Evento model)
+        public async Task<IActionResult> Post(EventoDto model)
         {
             try
             {
@@ -96,12 +97,12 @@ namespace ProEventos.API.Controllers
         }
 
         [HttpPut("{id}")]
-        public async Task<IActionResult> Put(int id, Evento model)
+        public async Task<IActionResult> Put(int id, EventoDto model)
         {
             try
             {
                 var eventos = await _eventoService.UpdateEvento(id, model);
-                if (eventos == null) return BadRequest("Erro ao tentar atualizar evento.");
+                if (eventos == null) return NoContent();
 
                 return Ok(eventos);
             }
@@ -118,9 +119,12 @@ namespace ProEventos.API.Controllers
         {
             try
             {
+                var eventos = await _eventoService.GetEventoByIdAsync(id, true);
+                if (eventos == null) return NoContent();
+
                 return await _eventoService.DeleteEvento(id) ? 
                         Ok("Deletado") :                
-                        BadRequest("Evento não deletado.");
+                        throw new Exception("Ocorreu um problema não específico ao tentar deletar Evento");
             }
             catch (Exception ex)
             {
